src/main/java/sap/service/MemberServiceImpl.java
package sap.service;

import sap.domain.Member;
import sap.repository.MemberRepository;

import java.util.Optional;

public class MemberServiceImpl implements MemberService{
    private final MemberRepository memberRepository;

    public MemberServiceImpl(MemberRepository memberRepository) {
        this.memberRepository = memberRepository;
    }

    @Override
    public Long join(Member member) {
        memberRepository.save(member);
        return member.getMember_id();
    }

    @Override
    public Optional<Member> findMember(Long memberId) {
        return memberRepository.findById(memberId);
    }
}
